MQ configure 코드 이동
