+++ consumer/src/main/java/com/sippet/consumer/MQConsumerConfiguration.java
@@ -1,8 +1,42 @@
 package com.sippet.consumer;
 
+import org.springframework.amqp.core.MessageListener;
+import org.springframework.amqp.core.Queue;
+import org.springframework.amqp.rabbit.connection.CachingConnectionFactory;
+import org.springframework.amqp.rabbit.connection.ConnectionFactory;
+import org.springframework.amqp.rabbit.listener.SimpleMessageListenerContainer;
+import org.springframework.amqp.rabbit.listener.adapter.MessageListenerAdapter;
+import org.springframework.amqp.support.converter.SimpleMessageConverter;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class MQConsumerConfiguration {
+    final static String queueName = "testMQ";
 
+    @Bean
+    ConnectionFactory connectionFactory() {
+        return new CachingConnectionFactory("localhost", 5672);
+    }
+
+    @Bean
+    Queue queue() {
+        return new Queue(queueName, false);
+    }
+
+    @Bean
+    MessageListener messageListener() {
+        return new MessageListenerAdapter(new Consumer(), new SimpleMessageConverter());
+    }
+
+    @Bean
+    public SimpleMessageListenerContainer container() {
+        SimpleMessageListenerContainer container = new SimpleMessageListenerContainer();
+        container.setConnectionFactory(connectionFactory());
+        container.setQueueNames(queueName);
+        container.setAutoStartup(true);
+        container.setConcurrentConsumers(2);
+        container.setMessageListener(messageListener());
+        return container;
+    }
 }
